create order with no currentuser

+++ Application/Services/IProductServices.cs
@@ -11,10 +11,10 @@ namespace Application.Services
     public interface IProductServices
     {
         Task<ResponseDTO> GetAllProductAsync(int pageIndex, int pageSize);
-        Task<ResponseDTO> GetProductByIdAsync(int productId);
+        Task<ResponseDTO> GetProductByIdAsync(long productId);
         Task<ResponseDTO> GetProductByNameAsync(string productName, int pageIndex, int pageSize);
         Task<ResponseDTO> CreateProductAsync(CreateProductDTO request);
-        Task<ResponseDTO> UpdateProductById(int productId, CreateProductDTO request);
-        Task<ResponseDTO> ChangeProductStatusById(int productId);
+        Task<ResponseDTO> UpdateProductById(long productId, CreateProductDTO request);
+        Task<ResponseDTO> ChangeProductStatusById(long productId);
     }
 }
