+++ Infrastructure/Repositories/IProductRepository.cs
@@ -9,7 +9,7 @@ namespace Infrastructure.Repositories
         Task<IEnumerable<Product>> GetPagedAsync(int pageIndex, int pageSize);
         Task<int> CountByNameAsync(string productName);
         Task<List<Product>> GetPagedByNameAsync(string productName, int pageIndex, int pageSize);
-        Task<Product?> GetProductById(int productId);
+        Task<Product?> GetProductById(long productId);
         Task<List<Product?>> GetProductByNameAsync(string productName);
         Task<bool> ExistsByNameAsync(string name);
     }
